+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `ElButton` `outline` and `destructive` variants
+
+### Changed
+
+- `ElButton` icons use `sm` through `md` buttons and `md` on `lg`
+
 ## [1.0.1] - 2026-08-22
 
 ### Changed